[#1087] Refactored body of the CommitsTable component.
